+++ upp/classes/preprocessing_config.py
@@ -20,6 +20,7 @@ from yamlinclude import YamlIncludeConstructor
 
 from upp import __version__
 from upp.classes.components import Components
+from upp.classes.plotting_config import PlottingConfig
 from upp.classes.resampling_config import ResamplingConfig
 from upp.classes.reweight_config import ReweightConfig
 from upp.classes.variable_config import VariableConfig
@@ -225,6 +226,10 @@ class PreprocessingConfig:
             if "reweighting" in self.config
             else None
         )
+        self.plotting = PlottingConfig(**self.config.get("plotting", {}))
+        if self.plotting.num_jets_plotting is None:
+            self.plotting.num_jets_plotting = self.num_jets_estimate_plotting
+
         # reproducibility
         try:
             self.git_hash = get_git_hash(Path(__file__).parent)